Example Projects for RA/FSP

+++ example_projects/ek_ra6m4/wifi/wifi_ek_ra6m4_ep/e2studio/src/wifi_ep.h
@@ -103,7 +103,7 @@ typedef struct
     char ssid[BUFF_LEN];
     char pwd[BUFF_LEN];
     uint32_t security;
-    uint8_t ip_addr_device[4];
+    uint32_t ip_addr_device[4];
     fsp_err_t status;
 }WiFiParameters_t;
